fix(blog): address round-4 PR #99 review findings
Layering + bundle
=================
- safeFeaturedImageUrl moved out of src/lib/absolute-url.ts into its
  own module (src/lib/featured-image-url.ts). absolute-url.ts is back
  to being a pure URL primitive (SITE_ORIGIN + canonicalUrl) with
  zero schema/zod dependencies — no more backwards layering
  (schemas → featured-image-url → absolute-url is the correct
  direction) and no more risk of dragging zod + the whole schemas
  module into client components that just import canonicalUrl.

- siteConfig.url and siteConfig.author.url pinned to SITE_ORIGIN.
  The env-aware `resolvedSiteUrl` (NEXT_PUBLIC_SITE_URL || localhost)
  fallback is gone — the very mechanism the 2026-05-07 incident
  comment described as the bug source. After this commit there is
  ONE source of site origin (SITE_ORIGIN); siteConfig still owns
  non-URL identity (name, description, social links).

Schema cleanups
===============
- New `nullishUrl(max=2048)` sibling helper to `nullishText`. The
  canonicalUrl field in blogPostBaseShape was hand-rolling the same
  3-branch union (literal('') transform → null, url, null); now uses
  the helper. Same `'' → null` parse-time guarantee.

- `nullishText` and `nullishUrl` now exported so future schemas
  (admin projects, contact-submission overrides) inherit the same
  HTML-form-friendly nullable-text contract by composition instead
  of each schema re-inventing the union.

- `nullishText`'s `message ? z.string().max(max, message) : z.string()
  .max(max)` ternary collapsed to `z.string().max(max, message)` —
  Zod's max() already accepts `undefined` for message.

- blogPostBaseShape preamble comment refreshed to point at the
  nullishText/nullishUrl mechanism (was still describing `.nullish()`).

Small polish
============
- safeFeaturedImageUrl: removed the optional 3rd-arg branch; both
  callers always passed 'Blog Post', so the parameter now defaults
  to the BLOG_OG_SUBTITLE constant inside the helper. Call sites
  drop the literal string.

- seo-provider.tsx: dropped the `canonicalUrl as buildCanonicalUrl`
  import alias (which existed only to dodge shadowing). Local var
  renamed to `pageCanonicalUrl` — `grep canonicalUrl` now finds this
  file like every other caller.

Verified
========
- bun run ci:quick: 468 files lint + typecheck clean
- bunx vitest run: 1003/1003 pass

[hudsor01]
