+++ src/lib/absolute-url.ts
@@ -1,11 +1,16 @@
-import { featuredImageSchema } from '@/lib/schemas'
-
 /**
  * Site origin used in SEO surfaces: canonical URLs, OG cards, JSON-LD
  * `image`/`url`, sitemap entries. Hardcoded to the production origin
  * because Search Console expects the production URL in `<link
  * rel="canonical">` even on preview deploys — varying by environment
  * would break the canonical contract.
+ *
+ * This module is deliberately a leaf — pure URL primitives with NO
+ * schema or validation imports. The composed validate-then-build
+ * helper (`safeFeaturedImageUrl`) lives in
+ * `src/lib/featured-image-url.ts` so adding it never inverts the
+ * dependency layering or drags `zod`/schemas into any new client
+ * surface that just needs `SITE_ORIGIN`/`canonicalUrl()`.
  */
 export const SITE_ORIGIN = 'https://richardwhudsonjr.com'
 
@@ -28,28 +33,4 @@ export const SITE_ORIGIN = 'https://richardwhudsonjr.com'
  */
 export function canonicalUrl(pathOrUrl: string): string {
   return /^https?:\/\//i.test(pathOrUrl) ? pathOrUrl : `${SITE_ORIGIN}${pathOrUrl}`
-}
-
-/**
- * Return a canonical absolute URL for a stored `featuredImage` value,
- * falling back to the branded `/api/og?title=…` card if the stored
- * value fails `featuredImageSchema` validation. Used by sitemap +
- * BlogPostJsonLd + any future surface that emits a featured image —
- * legacy/imported rows never went through the write-path validator,
- * so re-validating at read time is the only thing standing between a
- * `//evil.com/x` value and Google indexing it.
- *
- * Lives here (not in schemas.ts) so callers don't have to import the
- * schema directly to get the safe URL.
- */
-export function safeFeaturedImageUrl(
-  stored: string | null | undefined,
-  fallbackTitle: string,
-  fallbackSubtitle?: string
-): string {
-  const parsed = stored ? featuredImageSchema.safeParse(stored) : null
-  if (parsed?.success && parsed.data) return canonicalUrl(parsed.data)
-  const params = new URLSearchParams({ title: fallbackTitle })
-  if (fallbackSubtitle) params.set('subtitle', fallbackSubtitle)
-  return canonicalUrl(`/api/og?${params.toString()}`)
 }
